+++ library/src/main/java/com/wultra/android/mtokensdk/operation/OperationsService.kt
@@ -255,7 +255,7 @@ class OperationsService {
 
                 override fun onTimeSynchronizationFailed(t: Throwable) {
                     WMTLogger.e("Proximity check: time synchronization failed: ${t.message}")
-                    callback(Result.failure(ApiErrorException(t)))
+                    callback(Result.failure(t))
                 }
             })
         }
@@ -268,6 +268,8 @@ class OperationsService {
      */
     private fun adjustProximityCheckData(proximityCheck: ProximityCheck): ProximityCheckData {
         val timeService = powerAuthSDK.timeSynchronizationService
+        check(timeService.isTimeSynchronized) { "adjustProximityCheckData called before time synchronization" }
+
         val localTimeAdjustment = timeService.localTimeAdjustment
 
         val originalReceived = proximityCheck.timestampReceived